Sets cpd sequence number explicitly.
When running this machinery in a loop sans batch api, cpd children
are created in the order that they are found in the original obj.

While the batch preprocessor still creates batch objects in this
order, the execution of the batch runs in dictionary order (pids)
which can result in objects 98,99 appearing after objects 100,101.